add labeling support for exported scrape jobs
The previous work to collect scrape jobs (#304) was a good start, but
it really becomes powerful when you can select certain labels to apply
on the exporters.

For example, I use this to export the node roles' in the prometheus
metrics, which allows me to regroup results by role instead of by node.