fix Archlinux
This commit fixes the Archlinux tests. Most notably the default for
`install_method` is reset to the default for Archlinux. This is because
most exporters actually don't have a package for Archlinux. The changes
are somwhat backwards incompatible, but I would argue that previously the
module was broken for Archlinux. So I consider this a bugfix instead.