Change openldap_exporter to a new fork
Users are advised to check their usage particulary if overriding the
version of the archive.

The original repository has been archived, this fork is regularly
updating the dependencies to fix security issues.

This introduces a change in the logic to manage the download, to
simplify greatly the code, the support for the original repository was
removed.

For convenience, the name of the service has not been changed, this
removes the need to perform a cleanup on the agents after updating the
module.